Improvement Upgrade Slop Gem
Upgrade slop gem on pry-remote to version 4ish.

Using pry-remote on projects like Gruf Framework is
throwing errors because slop gem is really old and not
compatible. We upgraded the version in one fork that
I created and it works.

Updating the gem reference on pry-remote.gemspec file
bumping the version to ~>4.0